pyproject.toml
[project]
name = "opensable"
version = "0.1.0-beta"
description = "Your personal AI that actually does things – autonomous, local, and yours forever"
authors = [{name = "IdeoaLabs", email = "contact@ideolabs.ai"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
keywords = ["ai", "agent", "agi", "autonomous", "llm", "telegram", "chatbot"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = []

[tool.setuptools]
packages = ["opensable", "opensable.core", "opensable.interfaces", "opensable.skills"]

[project.optional-dependencies]
core = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "langgraph>=0.0.40",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "ollama>=0.1.6",
    "chromadb>=0.4.22",
    "tiktoken>=0.5.2",
    "aiogram>=3.3.0",
    "python-telegram-bot>=20.7",
    "telethon>=1.34.0",
    "discord.py>=2.3.2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "httpx>=0.26.0",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
    "click>=8.1.7",
    "psutil>=5.9.0",
    "RestrictedPython>=6.2",
    "aiofiles>=23.2.1",
]

voice = [
    "faster-whisper>=0.10.0",
    "piper-tts>=1.2.0",
    "pyttsx3>=2.90",
]

vision = [
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "pillow>=10.1.0",
    "ultralytics>=8.0.200",
    "easyocr>=1.7.0",
]

sync = [
    "websockets>=12.0",
    "cryptography>=41.0.7",
]

automation = [
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "selenium>=4.16.0",
    "fake-useragent>=1.4.0",
]

database = [
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "aiomysql>=0.2.0",
    "motor>=3.3.0",
    "redis[asyncio]>=5.0.0",
]

integrations = [
    "discord.py>=2.3.2",
    "slack-bolt>=1.18.0",
    "google-api-python-client>=2.110.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "twilio>=8.10.0",
]

monitoring = [
    "prometheus-client>=0.19.0",
]

dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.2",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.12.1",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/IdeoaLabs/Open-Sable"
Documentation = "https://github.com/IdeoaLabs/Open-Sable/tree/master/docs"
Repository = "https://github.com/IdeoaLabs/Open-Sable"
Issues = "https://github.com/IdeoaLabs/Open-Sable/issues"

[project.scripts]
opensable = "opensable.__main__:main"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
